+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/Accessibility.kt
@@ -1,5 +1,0 @@
-package com.stripe.android.ui.core
-
-internal fun String.asIndividualDigits(): String {
-    return toCharArray().joinToString(separator = " ")
-}