+++ paymentsheet-example/src/androidTest/java/com/stripe/android/test/core/FieldPopulator.kt
@@ -18,6 +18,7 @@ import com.stripe.android.paymentsheet.example.playground.settings.DefaultBillin
 import com.stripe.android.test.core.ui.Selectors
 import com.stripe.android.ui.core.elements.TranslationId
 import com.stripe.android.ui.core.elements.formatExpirationDateForAccessibility
+import com.stripe.android.uicore.utils.asIndividualDigits
 import com.stripe.android.core.R as CoreR
 
 internal class FieldPopulator(
@@ -142,10 +143,10 @@ internal class FieldPopulator(
     private fun validateZip() {
         if (usesZip()) {
             selectors.getZip()
-                .ifExistsAssertContentDescriptionEquals(values.zip)
+                .ifExistsAssertContentDescriptionEquals(values.zip.asIndividualDigits())
         } else {
             selectors.getPostalCode()
-                .ifExistsAssertContentDescriptionEquals(values.zip)
+                .ifExistsAssertContentDescriptionEquals(values.zip.asIndividualDigits())
         }
     }